fix: create testingFormat & findingFile function and add to validation func
