+++ src/validations/index.js
@@ -1,4 +1,6 @@
 const isIncluded = require('./isIncluded');
 const isDuplicated = require('./isDuplicated');
+const testingFormat = require('./testingFormat');
+const findingFile = require('./findingFile');
 
-module.exports = { isIncluded, isDuplicated };
+module.exports = { isIncluded, isDuplicated, testingFormat, findingFile };
